Implemented reconcile from Shopify maintenance utility

+++ hinundhergestellt-vendor-shopify-api/src/main/kotlin/de/hinundhergestellt/jhuh/vendors/shopify/client/ShopifyWeight.kt
@@ -3,10 +3,11 @@ package de.hinundhergestellt.jhuh.vendors.shopify.client
 import de.hinundhergestellt.jhuh.vendors.shopify.graphql.types.Weight
 import de.hinundhergestellt.jhuh.vendors.shopify.graphql.types.WeightInput
 import de.hinundhergestellt.jhuh.vendors.shopify.graphql.types.WeightUnit
+import java.math.BigDecimal
 
 data class ShopifyWeight(
     val unit: WeightUnit,
-    val value: Double
+    val value: BigDecimal
 ) {
     internal constructor(weight: Weight) : this(
         unit = weight.unit,
